Add search over diagnosis
This may not the efficient way to search but will work for now.
Down the road, we might have to change the way we search now

I have added querydsl which seems really interesting to me, so putting
this as an experiment, if things work well, we might keep this.