Add delete option in picture information
Ideally we are not deleting anything, but in case user wants to remove,
they can remove the file/image from the picture information page.
User won't be able to see, but we are not removing from the database.

Note: the way we have developed the file storage, isn't actually ideal,
down the road, we have to think about a better alternative and the soon we do,
better, otherwise migration would be a real pain like hell.